Revert "Seweryn"
