+++ Assets/Scripts/CameraMove.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CameraMove : MonoBehaviour
+{
+
+    public Transform CameraPos;
+
+    // Update is called once per frame
+    void Update()
+    {
+        transform.position = CameraPos.position;
+    }
+}
