+++ Assets/Scripts/CameraMove.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7920646f6572cea42a1ea79b1099d7eb
+guid: 74eec676223c67740ae6dc847cb5224a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
