+++ Assets/Scripts/FixCamera.cs
