--- Assets/Scripts/Interaction.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a592a80607c016f47917dcbd19d4ee68
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
