--- Assets/Scripts/Interaction/PressButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d946178c811bed849aa3b3c82066eab9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
