--- Assets/Scripts/PLayer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 80dc3b6de90d35145ae233712c450755
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
