--- Assets/Scripts/PLayer/PlayerAtributes.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e950760887ed2bf4a8426f2955300cf5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
