--- Assets/Scripts/buttons/AddHealth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0a561561bb5e59c4bbf39c052cd8b7f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
