--- Assets/Scripts/buttons/AddStamina.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60544a270f40371478685244d2436172
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
