--- Assets/sprites2D/Materials/ang_logo.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8ca64504940fdd74c84ff9f9692dd2ee
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
