+++ ProjectSettings/TagManager.asset
@@ -3,8 +3,7 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags:
-  - UI_point1
+  tags: []
   layers:
   - Default
   - TransparentFX
